+++ src/client/jupyter/jupyterNotInstalledNotificationHelper.ts
